implemented new money object

+++ app/src/main/java/com/example/agenda/domain/entities/Bank.kt
@@ -1,6 +1,5 @@
 package com.example.agenda.domain.entities
 
-import MONEY
 import com.example.agenda.app.common.Entity
 import io.objectbox.annotation.Entity as E
 import io.objectbox.annotation.Id
@@ -8,11 +7,11 @@ import io.objectbox.annotation.Id
 @E
 data class Bank(
     val name: String,
-    var balance: MONEY,
+    var balance: Double,
     override var id: String?,
     @Id
     var _id: Long = 0,
     override var created_at: Long?,
     override var updated_at: Long?,
-    val credit: MONEY?,
+    val credit: Double?,
 ) : Entity {init {createMetadata()}}
